Move animated sprite logic directly into cSprite
